Corrected rescopedEvent handler logic
Per https://answers.atlassian.com/questions/239988 and the discussion
on this pull request (#12), we should only send the pull request event
when there are changes on the "from" side of the pull request. Otherwise
events are potentially sent for every push to the destination repo.
This also forces the update of refs in the destination repo.